Backend refactor: extract chat service layer from routes
- New app/services/chat_service.py owns the ConnectionManager, Anthropic streaming, history, rate limits, and prompt caching
- chat_routes.py is now a thin layer: status endpoint, frame parsing, delegation
- Tests import from the service module; backend README architecture updated
